Update fuelrats.ini
Put a condition on the removal of semicolons... 80129aa ended up adding an equals in front of the CMDR name displayed in the listbox and the hashtable data. The actual hashtable record was properly named and would have matched a clear, but the listbox line would have remained.

Also put the lovely • as a delimiter for the rsig announce lines, to match the clear lines.