Update fuelrats.ini
Add a space to system names that begin with a number so they can be colored properly on display in the listbox. Combining a color code directly with a string starting with a number will bork the color application. Still applies to CMDR names starting with numbers - need to fix.