Update fuelrats.ini
Re: Notes on e9eb976
On Ratsignal, coloring of CMDR names beginning with a number was working, but was broken on !clear because the signal announce was taking data from the ratsignal_active hashtable, and the clear announce was taking data from a variable. Updated the !clear announce to read from the ratsignal_cleared hashtable, and moved the ratsignal_cleared hashtable update to occur before the announce of !clear.